requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
urllib3>=2.0.0
pytest>=7.4.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic-settings>=2.6.0
httpx>=0.27.2
tenacity>=9.0.0
cachetools>=5.3.3
trio>=0.26.0
mcp>=1.0.0
fastmcp>=2.0.0

